+++ packages/json-rpc/sm-provider/src/sm-provider.ts
@@ -1,4 +1,4 @@
-import type { Chain } from "smoldot"
+import type { Chain } from "@polkadot-api/smoldot"
 import type { JsonRpcProvider } from "@polkadot-api/json-rpc-provider"
 import { getSyncProvider } from "@polkadot-api/json-rpc-provider-proxy"